优化下载脚本，支持读取本地的wac文件

+++ scripts/create-project/server.sh
@@ -3,7 +3,12 @@
 # cd /your_project
 
 # Get codes
-curl -L https://github.com/yuri2peter/web-aircraft-carrier/archive/refs/heads/main.zip -o wac.zip
+if [ -e "wac.zip" ]; then
+  echo "File wac.zip exists."
+else
+  echo "Downloading file wac.zip ..."
+  curl -L https://github.com/yuri2peter/web-aircraft-carrier/archive/refs/heads/main.zip -o wac.zip
+fi
 
 # Decompression
 unzip wac.zip
